make all opm modules compile fine with the vanilla master versions
this requires to replace the HAVE_CONFIG_H compile definition by
HAVE_CONFIG_H=1 and to emulate the sources_hook() hack of the current
OPM build system for some modules.

Note that most of the sources_hook() for opm-core currently does
nothing because it deals with files that do not exist anymore or check
for hard dependencies. I took the liberty of not replicating that
functionality.