add a BUILD_APPLICATIONS cmake switch
this allows to shave off some seconds of built time if only downstream
modules are of interest. the switch works completely analogous to
BUILD_TESTING: if it is off, the application binaries are not build by
default, but building can be forced by `make $APPNAME`. In addition to
the behaviour for test programs, the applications are now only marked
for installation if BUILD_APPLICATIONS is true.