add support for opm-parser
the custom build system of opm-parser recently has been removed, so we
need to pick up the slack. this is not very pretty -- mainly because
the code generator needs to be called during the build and because the
opm-parser file locations are quite non OPM-standard --, but for now
it seems to work well enough until the next module layout refactoring
happens.