opm-parser: deal with the "ExtraTests"
what the hell is the difference of those to "normal" tests? This patch
is a simple solution which piggy-backs on the default opm-parser build
system, i.e. "make build_tests" does not work for these tests, but IMO
it is not worth to get into a whack-a-mole game because of this...