fix conflicts with dune-alugrid with FindZOLTAN.cmake
FindZOLTAN.cmake of dune-alugrid seems to be buggy as it does not find
my system-installed ZOLTAN libraries. For opm-grid itself this did not
bite because the dune buildsystem seems to make sure that a module's
local cmake directory comes first, but opm modules that depend on
opm-grid did not detect ZOLTAN properly.

the work-around is that opm-grid ensures that its module search path
precedes the one of dune-alugrid in depending modules.