make it work with CMake 3.0
this allows to use distributions that are based on debian jessie which
features cmake 3.0.2. the main problem is that this cmake version
seems to be alergic to specifing to link to a given library multiple
times. This requires a slightly modified version of the FindBoost
cmake module that is shipped with cmake 3.10.